mudança para o reativo

+++ src/main/java/dev/luizpais/bank/ContaCorrente.java
@@ -1,11 +1,8 @@
 package dev.luizpais.bank;
 
-import io.quarkus.hibernate.orm.panache.PanacheEntityBase;
+import io.quarkus.hibernate.reactive.panache.PanacheEntityBase;
 import jakarta.persistence.*;
 import lombok.*;
-
-import java.math.BigDecimal;
-import java.util.Date;
 
 @Setter
 @Getter
@@ -18,7 +15,7 @@ public class ContaCorrente extends PanacheEntityBase {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id", nullable = false)
-    private Integer id;
+    private Long id;
 
     @Column(name = "nome")
     private String nome;
